populateLemmaIndex when tab clicked
